fix(xp): heartbeat de tempo online quebrado por type inference (500)
RankingStorage.heartbeat usava LEAST($3, $2) com parametros sem tipo -> Postgres infere TEXT -> 'column minutes_online is of type integer but expression is of type text' -> HTTP 500 em TODA chamada de heartbeat. Resultado: user_online_time vazio globalmente e ZERO XP de tempo online desde sempre. Fix: cast explicito $3::int / $2::int. Os demais eventos de XP nunca passaram por aqui, por isso funcionavam.

+++ src/storages/RankingStorage.js
@@ -229,9 +229,9 @@ module.exports = {
        ),
        upd AS (
          INSERT INTO user_online_time (id_user, date, minutes_online)
-         VALUES ($1, CURRENT_DATE, LEAST($3, $2))
+         VALUES ($1, CURRENT_DATE, LEAST($3::int, $2::int))
          ON CONFLICT (id_user, date) DO UPDATE
-           SET minutes_online = LEAST(user_online_time.minutes_online + $3, $2)
+           SET minutes_online = LEAST(user_online_time.minutes_online + $3::int, $2::int)
          RETURNING minutes_online AS new
        )
        SELECT upd.new, COALESCE(prev.old, 0) AS old
